Fix "file_lines" issue in examples/user/user.java
