+++ examples/user/GetWebhookParseSettings.java
@@ -1,0 +1,21 @@
+//////////////////////////////////////////////////////////////////
+// Retrieve all parse settings
+// GET /user/webhooks/parse/settings
+
+
+public class Example {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.GET);
+      request.setEndpoint("user/webhooks/parse/settings");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
