+++ examples/user/README.md
@@ -1,0 +1,27 @@
+user examples:
+
+sendgrid-java/examples/user/CreateWebhookParseSettings.java
+sendgrid-java/examples/user/DeleteScheduleSend.java
+sendgrid-java/examples/user/DeleteWebhookParseSettings.java
+sendgrid-java/examples/user/GetCreditBalance.java
+sendgrid-java/examples/user/GetEmailAddress.java
+sendgrid-java/examples/user/GetEnforedTLS.java
+sendgrid-java/examples/user/GetProfile.java
+sendgrid-java/examples/user/GetScheduleSend.java
+sendgrid-java/examples/user/GetScheduleSendByBatchId.java
+sendgrid-java/examples/user/GetUserInfo.java
+sendgrid-java/examples/user/GetUsername.java
+sendgrid-java/examples/user/GetWebhookEventSettings.java
+sendgrid-java/examples/user/GetWebhookParseSettingByHostname.java
+sendgrid-java/examples/user/GetWebhookParseSettings.java
+sendgrid-java/examples/user/GetWebhookParseStatistics.java
+sendgrid-java/examples/user/PauseScheduleSend.java
+sendgrid-java/examples/user/TestWebhookEvent.java
+sendgrid-java/examples/user/UpatePassword.java
+sendgrid-java/examples/user/UpdateEmailAddress.java
+sendgrid-java/examples/user/UpdateEnforcedTLS.java
+sendgrid-java/examples/user/UpdateProfile.java
+sendgrid-java/examples/user/UpdateScheduleSend.java
+sendgrid-java/examples/user/UpdateUsername.java
+sendgrid-java/examples/user/UpdateWebhookEventSettings.java
+sendgrid-java/examples/user/UpdateWebhookParseSettings.java
