+++ examples/user/TestWebhookEvent.java
@@ -1,0 +1,22 @@
+//////////////////////////////////////////////////////////////////
+// Test Event Notification Settings
+// POST /user/webhooks/event/test
+
+
+public class Example {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.POST);
+      request.setEndpoint("user/webhooks/event/test");
+      request.setBody("{\"url\":\"url\"}");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
