+++ examples/user/UpatePassword.java
@@ -1,0 +1,22 @@
+//////////////////////////////////////////////////////////////////
+// Update your password
+// PUT /user/password
+
+
+public class Example {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.PUT);
+      request.setEndpoint("user/password");
+      request.setBody("{\"new_password\":\"new_password\",\"old_password\":\"old_password\"}");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
