+++ examples/user/UpdateUsername.java
@@ -1,0 +1,22 @@
+//////////////////////////////////////////////////////////////////
+// Update your username
+// PUT /user/username
+
+
+public class Example {
+  public static void main(String[] args) throws IOException {
+    try {
+      SendGrid sg = new SendGrid(System.getenv("SENDGRID_API_KEY"));
+      Request request = new Request();
+      request.setMethod(Method.PUT);
+      request.setBody("user/username");
+      request.setBody("{\"username\":\"test_username\"}");
+      Response response = sg.api(request);
+      System.out.println(response.getStatusCode());
+      System.out.println(response.getBody());
+      System.out.println(response.getHeaders());
+    } catch (IOException ex) {
+      throw ex;
+    }
+  }
+}
